requirement.txt
pandas
Flask
python-dotenv
tweepy
nltk
